AP_Landing: new param LAND_TYPE to split current behavior into AP_Landing_Slope.cpp
